move tests with shadow disk disabling to separate test suites
