+++ cloud/disk_manager/internal/pkg/facade/image_service_nrd_test/ya.make
@@ -1,0 +1,10 @@
+GO_TEST_FOR(cloud/disk_manager/internal/pkg/facade)
+
+SET_APPEND(RECIPE_ARGS --disk-agent-count 3)
+INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common.inc)
+
+GO_XTEST_SRCS(
+    image_service_nrd_test.go
+)
+
+END()
