+++ cloud/disk_manager/internal/pkg/facade/image_service_test/ya.make
@@ -3,7 +3,6 @@ GO_TEST_FOR(cloud/disk_manager/internal/pkg/facade)
 SET_APPEND(RECIPE_ARGS --generate-vmdk-image)
 SET_APPEND(RECIPE_ARGS --generate-big-raw-images)
 SET_APPEND(RECIPE_ARGS --encryption)
-SET_APPEND(RECIPE_ARGS --disk-agent-count 3)
 INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common.inc)
 
 GO_XTEST_SRCS(
